+++ testdata/Colfer.go
@@ -337,7 +337,7 @@ func (o *O) MarshalLen() (int, error) {
 	}
 
 	if v := o.T; !v.IsZero() {
-		if s := uint(v.Unix()); s < 1<<32 {
+		if s := uint64(v.Unix()); s < 1<<32 {
 			l += 9
 		} else {
 			l += 13
